refactor: расщепление api.py на роутеры + JWT-секрет обязателен + Pinia кэш пагинации
- src/web/api.py: 1394 строк → 167 (фабрика приложения + SSE + health)
- src/web/deps.py: общие зависимости (JWT, CRM-клиент, пагинация, сериализаторы, модели)
- src/web/routers/: 7 модулей (auth, dashboard, orders, clients, products, export, users)
- WEB_SECRET теперь обязателен (RuntimeError вместо random fallback)
- web/src/stores/pagination.js: Pinia-кэш страниц заказов и клиентов (TTL 2 мин)
- Восстановление фильтров и страницы при навигации назад
- systemd: удалён tg-socks.service, упрощён groq-tunnel.service
