--- systemd/tg-socks.service
@@ -1,18 +1,0 @@
-[Unit]
-Description=SOCKS5 proxy for Telegram Bot API (BEEBOT)
-After=network-online.target
-Wants=network-online.target
-
-[Service]
-Type=simple
-User=hive
-WorkingDirectory=/home/hive/BEEBOT
-ExecStart=/home/hive/BEEBOT/.venv/bin/python tg_socks_proxy.py
-Restart=always
-RestartSec=5
-StandardOutput=journal
-StandardError=journal
-SyslogIdentifier=tg-socks
-
-[Install]
-WantedBy=multi-user.target
